Move mapValues into utils
